Fix #1353: aimodel/test filenames are confusing
